# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Author's Haven",
    "description": "A Social platform for the creative at heart.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://author-haven-stage.herokuapp.com/api/v1",
      "description": "Staging server"
    },
    {
      "url": "https://author-haven.herokuapp.com/api/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/articles/{articleId}": {
      "parameters": [
        {
          "name": "articleId",
          "in": "path",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "ID of the article",
          "required": true
        }
      ],
      "get": {
        "tags": ["Articles"],
        "summary": "This endpoint gets a single article",
        "operationId": "getArticle",
        "responses": {
          "200": {
            "description": "Return an article",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "400": {
            "description": "article fetch error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "article fetch error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Articles"],
        "summary": "This endpoint updates a single article",
        "operationId": "updateArticle",
        "responses": {
          "202": {
            "description": "Updated article",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleUpdateSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "update article error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Articles"],
        "summary": "This endpoint deletes single article",
        "operationId": "deleteArticle",
        "responses": {
          "202": {
            "description": "successful article delete operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "article delete operation not successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/articles/{articleId}/publish": {
      "parameters": [
        {
          "name": "articleId",
          "in": "path",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "ID of the article",
          "required": true
        }
      ],
      "patch": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Articles"],
        "summary": "Article publish endpoint",
        "description": "This action can be used to publish an article",
        "operationId": "publishArticle",
        "responses": {
          "200": {
            "description": "Failure Operation",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failure Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/articles/{articleId}/bookmark": {
      "parameters": [
        {
          "name": "articleId",
          "in": "path",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "ID of the article",
          "required": true
        }
      ],
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Bookmarks"],
        "summary": "Article bookmarking",
        "description": "This endpoint can be used to bookmark and unbookmark an article",
        "operationId": "bookmarkArticle",
        "responses": {
          "200": {
            "description": "Failure Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "201": {
            "description": "Successful Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failure Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Failure Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failure Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "articles/{articleId}/ratings": {
      "parameters": [
        {
          "name": "articleId",
          "in": "path",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "ID of the article",
          "required": true
        }
      ],
      "/ratings": null,
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Ratings"],
        "summary": "User article rating",
        "description": "This endpoint can be used to rate an article",
        "operationId": "rateArticle",
        "requestBody": {
          "description": "The request body should contain a rating field",
          "content": {
            "*/*": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {
                    "description": "Article rating value",
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful article rating update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "201": {
            "description": "successful article rating creation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Rate article",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/articles/{id}/escalate": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "ID of the article",
          "required": true
        }
      ],
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Reports"],
        "summary": "This endpoint can be used to report an article",
        "operationId": "reportArticle",
        "requestBody": {
          "description": "The request body should contain a description and category fields",
          "content": {
            "*/*": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "Article report description"
                  },
                  "category": {
                    "type": "string",
                    "description": "Article report category"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "article report success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "article report error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "published": {
            "type": "boolean",
            "default": "false"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "favouritesCount": {
            "type": "integer"
          },
          "readCount": {
            "type": "integer"
          },
          "totalReadTime": {
            "type": "integer",
            "default": 1
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ArticleUpdateSuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "data": {
            "type": [],
            "$ref": "#/components/schemas/Article"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}